split production.py into cobb_douglas.py, reorganize tests
