+++ gEconpy/model/block/cobb_douglas.py
