+++ tests/model/block/test_cobb_douglas.py
@@ -1,15 +1,3 @@
-"""Tests for the block dispatch system and CobbDouglasBlock.
-
-Phase 1 deliverable for the special blocks roadmap (audits/special_blocks/roadmap.md).
-
-Validates that:
-  * The parser dispatches to CobbDouglasBlock for canonical Cobb-Douglas
-    production blocks, and falls back to Block for everything else.
-  * Detection is conservative (no false positives on near-CD forms).
-  * The closed-form FOCs produce a policy function (T, R) identical to the
-    general Block path at machine epsilon.
-"""
-
 import gc
 
 import numpy as np
@@ -17,19 +5,13 @@ import pytest
 import sympy as sp
 
 from gEconpy import model_from_gcn
-from gEconpy.model.block import Block, dispatch_block
+from gEconpy.model.block import Block
 from gEconpy.model.block import registry as registry_mod
-from gEconpy.model.block.production import CobbDouglasBlock, _match_cobb_douglas_constraint
+from gEconpy.model.block.cobb_douglas import CobbDouglasBlock, _match_cobb_douglas_constraint
 from gEconpy.parser.loader import load_gcn_file
 from gEconpy.solvers.cycle_reduction import solve_policy_function_with_cycle_reduction
 
 RBC_PATH = "gEconpy/data/GCN Files/RBC.gcn"
-
-
-@pytest.fixture
-def disable_dispatch(monkeypatch):
-    """Force the registry to dispatch nothing — all blocks become base Block."""
-    monkeypatch.setattr(registry_mod, "_REGISTRY", [])
 
 
 class TestDispatchOnRBC:
@@ -50,9 +32,7 @@ class TestDispatchOnRBC:
         """The FIRM block's FOCs must be in closed-form Y/K and Y/L form, not the chain-rule expansion."""
         m = model_from_gcn(RBC_PATH, verbose=False)
         eq_strings = [str(eq) for eq in m.equations]
-        # eq6 must contain Y_t/K_t-1, NOT K_t-1**(alpha - 1).
         assert any("Y_t/K_t-1" in s for s in eq_strings), f"Expected closed-form Y/K FOC, got equations: {eq_strings}"
-        # No chain-rule K**(alpha - 1) form anywhere.
         assert not any("K_t-1**(alpha - 1)" in s for s in eq_strings), (
             f"Chain-rule expansion leaked through: {eq_strings}"
         )
@@ -82,12 +62,10 @@ class TestPolicyEquivalence:
 
     def test_rbc_policy_equivalence(self, monkeypatch):
         """Compute T, R with and without dispatch and compare element-wise."""
-        # Dispatched run
         m = model_from_gcn(RBC_PATH, verbose=False)
         A, B, C, D = m.linearize_model(verbose=False)
         T, R, _, _ = solve_policy_function_with_cycle_reduction(A, B, C, D, 1000, 1e-12, False)
 
-        # Disable dispatch and re-run
         monkeypatch.setattr(registry_mod, "_REGISTRY", [])
         gc.collect()
         m_b = model_from_gcn(RBC_PATH, verbose=False)
@@ -118,14 +96,13 @@ class TestDetectionConservatism:
     def test_rejects_non_cd_constraint(self):
         """Constraint that doesn't structurally match CD → no match."""
         Y, A, x1, x2 = sp.symbols("Y A x1 x2")
-        # Linear production — not CD
         constraints = {0: sp.Eq(Y, A * (x1 + x2))}
         assert _match_cobb_douglas_constraint(constraints) is None
 
     def test_rejects_exponent_zero(self):
         """Wild excludes a_i=0 (degenerate input collapses to constant)."""
         Y, A, x1, x2 = sp.symbols("Y A x1 x2")
-        constraints = {0: sp.Eq(Y, A * x1**0 * x2**1)}  # = A * x2
+        constraints = {0: sp.Eq(Y, A * x1**0 * x2**1)}
         assert _match_cobb_douglas_constraint(constraints) is None
 
     def test_accepts_crs_form(self):
@@ -137,11 +114,9 @@ class TestDetectionConservatism:
         assert m["Y"] == Y
         assert m["A"] == A
         assert m["idx"] == 3
-        # Inputs come back as a list of (symbol, exponent) pairs in some order;
-        # check both inputs are present and the exponents are recovered.
         inputs = dict(m["inputs"])
         assert set(inputs.keys()) == {x1, x2}
-        assert inputs[x1] + inputs[x2] == 1  # a1 + (1 - a1) = 1
+        assert inputs[x1] + inputs[x2] == 1
 
     def test_accepts_non_crs_form(self):
         """Decreasing returns to scale (a1 + a2 != 1) must still match — the closed-form FOC holds for any monomial."""
@@ -159,8 +134,8 @@ class TestDetectionConservatism:
     def test_accepts_arbitrary_arity(self, k):
         """Arity-agnostic: Y = A * prod(x_i^a_i for i in 1..k) must match for any k >= 1."""
         Y, A = sp.symbols("Y A")
-        xs = sp.symbols(f"x1:{k + 1}")  # x1, x2, ..., xk
-        as_ = sp.symbols(f"a1:{k + 1}")  # a1, a2, ..., ak
+        xs = sp.symbols(f"x1:{k + 1}")
+        as_ = sp.symbols(f"a1:{k + 1}")
         prod = A
         for x, a in zip(xs, as_, strict=True):
             prod = prod * x**a
